Support MDX files
MDX is a superset of Markdown that allows for JavaScript and React
components. This can be helpful if the author wants to include Interface
components that are not part of the standard spec (e.g. a Signature
Block to be signed by an Electronic-Signature service).

Since this library uses remark parser for markdown via
annotatedtext-remark, supporting MDX shouldn't be too much work because
Remark can also parse MDX via plug-ins.

Here, I duplicated the tests in case you wanted to keep MDX and Markdown
separate, though using Remark, they could be the same code (strip out
all the JavaScript before/after running LanguageTools).

Thanks for this awesome library!